Improve provision system
At the moment we had these provision alternatives:
- initial deployment based on base64 encoded source code.
- docker container interaction (kubectl copy/exec)

In this commit we remove the base64 mechanism and implement
a configMap and a values 'provisionPath' to carry any file
accessible to the helm chart.
Also, we implement an administration interface (additional
service listening on 8074 by default) which support GET for
touch and POST for creation of provision resources.

This admin interface will allow to refactor the testing
inside containers (pytest like or similar which support
HTTP/1).

The new admin.py will be launched in background on starter
script, and main mock server (mock.py) is still launched
in blocking mode to directly restart the container in case
of crash. So, only liveness probe for admin.py will be
actually needed. Also readiness probe is implemented for it.

- Updated README.md file
- Allow control verbosity of flask applications
- Generalize flask app building (abstraction to support
  admin app creation)
- Extend NOTES.txt on deployment, with new admin interface.
- Removed helm test
- Random generation of secrets for both flask applications.